Add automation for tagging the GitHub Action
Add additional automation for the tagging of the GitHub Action,
including both for the README.md on tagging of the release, as well as
for the action.yaml when changes are made to the main branch in
preparation for release.

This will mean that the main branch may not always be able to launch a
run, as the action.yaml may point to a container which has not been
build or released yet.

The execution of the action has been changed from docker to composite
due to the issues with using docker containers and ghcr.io as a source
without authentication.