Add initial code to check the title
Add the initial code which checks the title of the pull request,
including the code to fetch and manage the pull request data itself and
interface between the action module and the github module.

Proper handling of failures is not yet written, such as providing a
comment back to the pull request, but an error is provided if the
minimum is not yet met.